improve: move classes into subfolders
