+++ src/main/java/com/abc/account/Account.java
@@ -4,6 +4,8 @@ import java.util.List;
 
 import com.abc.account.transaction.ITransaction;
 import com.abc.account.transaction.Transaction;
+import com.abc.account.transaction.TransactionRepository;
+import com.abc.account.types.AccountType;
 
 /**
  *
