+++ src/main/java/com/abc/account/types/SavingsAccount.java
@@ -1,14 +1,16 @@
-package com.abc.account;
+package com.abc.account.types;
+
+import com.abc.account.Account;
 
 /**
  * 
  * @author Isabel Peters (isabel.rlpeters@googlemail.com)
  *
  */
-public class CheckingsAccount extends Account {
+public class SavingsAccount extends Account {
 
-	public CheckingsAccount() {
-		super(AccountType.CHECKING);
+	public SavingsAccount() {
+		super(AccountType.SAVINGS);
 	}
 
 	@Override
