+++ src/main/java/com/abc/bank/Bank.java
@@ -2,6 +2,8 @@ package com.abc.bank;
 
 import java.util.logging.Level;
 
+import com.abc.bank.exception.NoCustomerException;
+import com.abc.bank.repository.CustomerRepository;
 import com.abc.customer.Customer;
 import com.abc.customer.ICustomer;
 import com.sun.istack.internal.logging.Logger;
