+++ src/main/java/com/abc/bank/CustomerSummaryGenerator.java
@@ -1,5 +1,6 @@
 package com.abc.bank;
 
+import com.abc.bank.repository.CustomerRepository;
 import com.abc.customer.ICustomer;
 
 /**
