+++ src/main/java/com/abc/bank/repository/CustomerRepository.java
@@ -1,9 +1,10 @@
-package com.abc.bank;
+package com.abc.bank.repository;
 
 import java.util.Collections;
 import java.util.LinkedList;
 import java.util.List;
 
+import com.abc.bank.exception.NoCustomerException;
 import com.abc.customer.ICustomer;
 
 /**
