+++ src/test/java/com/abc/bank/BankTest.java
@@ -12,6 +12,8 @@ import java.util.List;
 import org.junit.Before;
 import org.junit.Test;
 
+import com.abc.bank.exception.NoCustomerException;
+import com.abc.bank.repository.CustomerRepository;
 import com.abc.customer.Customer;
 import com.abc.customer.ICustomer;
